chore: move asset selection components into correct directory structure
